VMS-9818: fixed aac audio encoder slowing down framerate, added Map<K, V> to buffered_stream_consumer.h,
 bug with StreamConsumerManager choosing the wrong VideoConsumer.

<==== Mercurial data ====>
Branch: webcam_plugin

+++ plugins/usb_cam/src/camera/video_stream.h
@@ -41,7 +41,7 @@ public:
     void removeFrameConsumer(const std::weak_ptr<FrameConsumer>& consumer);
 
     AVPixelFormat decoderPixelFormat() const;
-    int fps() const;
+    float fps() const;
 
     void updateFps();
     void updateBitrate();
